test(conn): move review regression tests into package test files
Relocate corner-case tests from *_review_test.go into pool_test.go,
balancer_test.go, and connections_state_test.go next to the code they cover.
